requirements.txt
requests
filemagic
memcached
